pyproject.toml
[project]
name = "gymnasium-experiments"
version = "0.1.0"
description = "A collection of gym-type evnvironment related scripts"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ffmpeg-python>=0.2.0",
    "gymnasium[box2d]>=0.29.1",
    "huggingface-hub>=0.25.0",
    "moviepy>=1.0.3",
    "nbqa>=1.9.0",
    "opencv-python>=4.10.0.84",
    "optuna>=4.0.0",
    "pydantic>=2.9.2",
    "pymysql>=1.1.1",
    "pytest>=8.3.3",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "rich>=13.8.1",
    "stable-baselines3>=2.3.2",
    "tensorboard>=2.17.1",
    "torch>=2.4.1",
    "tqdm>=4.66.5",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.6.6",

]

[tool.ruff.lint]
select = ["I"]

[tool.ruff]
line-length = 88
